refactor(dbModel): Appoiments dbModel
