+++ app/src/main/java/com/example/dogapp/view/adapter/AppointmentsAdapter.kt
@@ -5,8 +5,8 @@ import android.view.ViewGroup
 import androidx.navigation.NavController
 import androidx.recyclerview.widget.RecyclerView
 import com.example.dogapp.databinding.AppointmentItemBinding
-import com.example.dogapp.view.model.Appointment
 import com.example.dogapp.view.viewholder.AppointmentsViewHolder
+import model.Appointment
 
 class AppointmentsAdapter(private val appointmentsList:MutableList<Appointment>, private val navController: NavController):
     RecyclerView.Adapter<AppointmentsViewHolder>() {
@@ -21,7 +21,7 @@ class AppointmentsAdapter(private val appointmentsList:MutableList<Appointment>,
     }
 
     override fun onBindViewHolder(holder: AppointmentsViewHolder, position: Int) {
-        val date = appointmentsList[position]
-        holder.setDateItem(date)
+        val appointment = appointmentsList[position]
+        holder.setAppointmentItem(appointment)
     }
 }
