--- app/src/main/java/com/example/dogapp/view/model/Appointment.kt
@@ -1,17 +1,0 @@
-package com.example.dogapp.view.model
-
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-import com.google.gson.annotations.SerializedName
-import java.io.Serializable
-
-@Entity
-data class Appointment(
-    @PrimaryKey(autoGenerate = true)
-    val id: Int = 0,
-    val petName: String,
-    val symptom: String,
-    val petBreed: String,
-    @SerializedName("petImage")
-    val petImage: String
-): Serializable
