--- app/src/main/java/com/example/dogapp/view/utils/Constants.kt
@@ -1,5 +1,0 @@
-package com.example.dogapp.view.utils
-
-object Constants {
-    const val NAME_BD: String ="appointments.db"
-}
