+++ app/src/main/java/data/AppointmentsDB.kt
@@ -1,11 +1,11 @@
-package com.example.dogapp.view.data
+package data
 
 import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
-import com.example.dogapp.view.model.Appointment
-import com.example.dogapp.view.utils.Constants.NAME_BD
+import model.Appointment
+import utils.Constants.NAME_DB
 
 @Database(entities = [Appointment::class], version = 1)
 abstract class AppointmentsDB: RoomDatabase() {
@@ -16,7 +16,7 @@ abstract class AppointmentsDB: RoomDatabase() {
             return Room.databaseBuilder(
                 context.applicationContext,
                 AppointmentsDB::class.java,
-                NAME_BD
+                NAME_DB
             ).build()
         }
     }
