+++ app/src/main/java/model/Appointment.kt
@@ -2,12 +2,16 @@ package model
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import com.google.gson.annotations.SerializedName
+import java.io.Serializable
 
 @Entity
 data class Appointment(
     @PrimaryKey(autoGenerate = true)
-    val id: Int =0,
-    val petName: Int,
-    val symptoms: String,
+    val id: Int = 0,
+    val petName: String,
+    val symptom: String,
+    val petBreed: String,
+    @SerializedName("petImage")
     val petImage: String
-)
+): Serializable
